test: update checked_hf_specifications.jsonl and test_all_abstasks.py
