+++ tests/checked_hf_specifications.jsonl
@@ -151,3 +151,21 @@
 {"path":"mteb/amazon_massive_scenario","revision":"7d571f92784cd94a019292a1f45445077d0ef634","task_name":"MassiveScenarioClassification"}
 {"path":"mteb/amazon_massive_intent","revision":"31efe3c427b0bae9c22cbb560b8f15491cc6bed7","task_name":"MassiveIntentClassification"}
 {"path":"mteb/tatoeba-bitext-mining","revision":"9080400076fbadbb4c4dcb136ff4eddc40b42553","task_name":"Tatoeba"}
+{"path":"rbawden/DiaBLa","revision":"5345895c56a601afe1a98519ce3199be60a27dba","task_name":"DiaBlaBitextMining"}
+{"path":"jinaai/ger_da_lir","revision":"0bb47f1d73827e96964edb84dfe552f62f4fd5eb","task_name":"GerDaLIR"}
+{"path":"narrativeqa","revision":"2e643e7363944af1c33a652d1c87320d0871c4e4","task_name":"NarrativeQARetrieval"}
+{"path":"facebook/flores","revision":"2db78afdeaccaedc3b33a95442a4e55766887e17","task_name":"FloresClusteringS2S"}
+{"path":"jinaai/spanish_passage_retrieval","revision":"9cddf2ce5209ade52c2115ccfa00eb22c6d3a837","task_name":"SpanishPassageRetrievalS2P"}
+{"path":"AI-Sweden/SuperLim","revision":"7ebf0b4caa7b2ae39698a889de782c09e6f5ee56","name":"dalaj","task_name":"DalajClassification"}
+{"path":"strombergnlp/nordic_langid","revision":"e254179d18ab0165fdb6dbef91178266222bee2a","name":"10k","task_name":"NordicLangClassification"}
+{"path":"lyon-nlp/mteb-fr-retrieval-syntec-s2p","revision":"aa460cd4d177e6a3c04fcd2affd95e8243289033","task_name":"SyntecRetrieval"}
+{"path":"lyon-nlp/alloprof","revision":"2df7bee4080bedf2e97de3da6bd5c7bc9fc9c4d2","task_name":"AlloprofRetrieval"}
+{"path":"mteb/germanquad-retrieval","revision":"f5c87ae5a2e7a5106606314eef45255f03151bb3","task_name":"GermanQuAD-Retrieval"}
+{"path":"sbx/superlim-2","revision":"ef1661775d746e0844b299164773db733bdc0bf6","name":"swedn","task_name":"SwednClustering"}
+{"path":"mteb/norquad_retrieval","revision":"9dcfcdb2aa578dd178330d49bf564248935f7fbe","task_name":"NorQuadRetrieval"}
+{"path":"mteb/nfcorpus","revision":"ec0fa4fe99da2ff19ca1214b7966684033a58814","task_name":"NFCorpus"}
+{"path":"GEM/opusparcus","revision":"9e9b1f8ef51616073f47f306f7f47dd91663f86a","task_name":"OpusparcusPC"}
+{"path":"jinaai/spanish_passage_retrieval","revision":"9cddf2ce5209ade52c2115ccfa00eb22c6d3a837","task_name":"SpanishPassageRetrievalS2S"}
+{"path":"clarin-knext/fiqa-pl","revision":"2e535829717f8bf9dc829b7f911cc5bbd4e6608e","task_name":"FiQA-PL"}
+{"path":"clarin-knext/arguana-pl","revision":"63fc86750af76253e8c760fc9e534bbf24d260a2","task_name":"ArguAna-PL"}
+{"path":"clarin-knext/nfcorpus-pl","revision":"9a6f9567fda928260afed2de480d79c98bf0bec0","task_name":"NFCorpus-PL"}
